Revert "Init chatroom"
