--- firebase-debug.log
@@ -1,60 +1,0 @@
-[debug] [2020-05-20T08:42:13.777Z] ----------------------------------------------------------------------
-[debug] [2020-05-20T08:42:13.780Z] Command:       /usr/local/bin/firebase /Users/iyor/.cache/firebase/tools/lib/node_modules/firebase-tools/lib/bin/firebase deploy
-[debug] [2020-05-20T08:42:13.780Z] CLI Version:   7.16.2
-[debug] [2020-05-20T08:42:13.780Z] Platform:      darwin
-[debug] [2020-05-20T08:42:13.780Z] Node Version:  v12.16.1
-[debug] [2020-05-20T08:42:13.782Z] Time:          Wed May 20 2020 03:42:13 GMT-0500 (Central Daylight Time)
-[debug] [2020-05-20T08:42:13.782Z] ----------------------------------------------------------------------
-[debug] [2020-05-20T08:42:13.782Z] 
-[debug] [2020-05-20T08:42:13.824Z] > command requires scopes: ["email","openid","https://www.googleapis.com/auth/cloudplatformprojects.readonly","https://www.googleapis.com/auth/firebase","https://www.googleapis.com/auth/cloud-platform"]
-[debug] [2020-05-20T08:42:13.825Z] > authorizing via signed-in user
-[debug] [2020-05-20T08:42:13.826Z] [iam] checking project checker-io for permissions ["cloudfunctions.functions.create","cloudfunctions.functions.delete","cloudfunctions.functions.get","cloudfunctions.functions.list","cloudfunctions.functions.update","cloudfunctions.operations.get","firebase.projects.get"]
-[debug] [2020-05-20T08:42:13.848Z] >>> HTTP REQUEST POST https://cloudresourcemanager.googleapis.com/v1/projects/checker-io:testIamPermissions  
-
-[debug] [2020-05-20T08:42:14.203Z] <<< HTTP RESPONSE 200
-[info] 
-[info] === Deploying to 'checker-io'...
-[info] 
-[info] i  deploying functions
-[info] Running command: npm --prefix "$RESOURCE_DIR" run lint
-[info] Running command: npm --prefix "$RESOURCE_DIR" run build
-[info] ✔  functions: Finished running predeploy script.
-[debug] [2020-05-20T08:42:33.296Z] > [functions] package.json contents: {
-  "name": "functions",
-  "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "build": "tsc",
-    "serve": "npm run build && firebase emulators:start --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "engines": {
-    "node": "8"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "dotenv": "^8.2.0",
-    "firebase": "^7.14.4",
-    "firebase-admin": "^8.12.1",
-    "firebase-functions": "^3.6.1",
-    "generate-password": "^1.5.1"
-  },
-  "devDependencies": {
-    "tslint": "^5.12.0",
-    "typescript": "^3.2.2",
-    "firebase-functions-test": "^0.1.6"
-  },
-  "private": true
-}
-[info] i  functions: ensuring necessary APIs are enabled...
-[debug] [2020-05-20T08:42:33.297Z] >>> HTTP REQUEST GET https://serviceusage.googleapis.com/v1/projects/checker-io/services/cloudfunctions.googleapis.com  
- 
-[debug] [2020-05-20T08:42:33.298Z] >>> HTTP REQUEST GET https://serviceusage.googleapis.com/v1/projects/checker-io/services/runtimeconfig.googleapis.com  
- 
-[debug] [2020-05-20T08:42:34.190Z] <<< HTTP RESPONSE 200
-[debug] [2020-05-20T08:42:34.335Z] <<< HTTP RESPONSE 200
-[info] ✔  functions: all necessary APIs are enabled
-[debug] [2020-05-20T08:42:34.337Z] >>> HTTP REQUEST GET https://firebase.googleapis.com/v1beta1/projects/checker-io/adminSdkConfig  
-
