--- functions/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "functions",
-  "scripts": {
-    "lint": "tslint --project tsconfig.json",
-    "build": "tsc",
-    "serve": "npm run build && firebase emulators:start --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "engines": {
-    "node": "8"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "dotenv": "^8.2.0",
-    "firebase": "^7.14.4",
-    "firebase-admin": "^8.12.1",
-    "firebase-functions": "^3.6.1",
-    "generate-password": "^1.5.1"
-  },
-  "devDependencies": {
-    "tslint": "^5.12.0",
-    "typescript": "^3.2.2",
-    "firebase-functions-test": "^0.1.6"
-  },
-  "private": true
-}
