--- public/firebaseConfig.json
@@ -1,10 +1,0 @@
-var data = {
-    "apiKey" : "AIzaSyBSStZGTofCfLQDJeGFYnZgYs1sUIW0GhU",
-    "authDomain" : "checker-io.firebaseapp.com",
-    "databaseURL" : "https://checker-io.firebaseio.com",
-    "projectId" : "checker-io",
-    "storageBucket" : "checker-io.appspot.com",
-    "messagingSenderId" : "119022070241",
-    "appId" : "1:119022070241:web:4cab5ec53a70bf508f0aa0",
-    "measurementId" : "G-Y4XHBBVNC5"
-}
