+++ src/Scripts/Board.js
@@ -1,44 +1,38 @@
-import { Piece } from '../Scripts/Piece.js';
-import { Position } from '../Scripts/Position.js';
+import { Piece } from './Piece.js';
+import { Position } from './Position.js';
 
 export class Board {
 
-    constructor(state) {
+    constructor() {
 
-        if(!arguments.length) {
+        this.board = [];
 
-            this.board = [];
+        let i, j, k = 0;
 
-            let i, j, k = 0;
+        for(i = 0; i < 8; i++) {
 
-            for(i = 0; i < 8; i++) {
+            this.board[i] = [];
+        
+            for(j = 0; j < 8; j++) {
+        
+                let even = (i % 2 == 0) && (j % 2 == 0);
+                let odd = (i % 2 != 0) && (j % 2 != 0);
+        
+                if(even || odd || i == 3 || i == 4) {
+                
+                    this.board[i][j] = null;
 
-                this.board[i] = [];
-            
-                for(j = 0; j < 8; j++) {
-            
-                    let even = (i % 2 == 0) && (j % 2 == 0);
-                    let odd = (i % 2 != 0) && (j % 2 != 0);
-            
-                    if(even || odd || i == 3 || i == 4) {
-                    
-                        this.board[i][j] = null;
+                } else  {
 
-                    } else  {
+                    if(0 <= i && i <= 2)
+                        this.board[i][j] = new Piece(i, j, 'U', k);
+                        
+                    else 
+                        this.board[i][j] = new Piece(i, j, 'D', k);
 
-                        if(0 <= i && i <= 2)
-                            this.board[i][j] = new Piece(i, j, 'U', k, null);
-                            
-                        else 
-                            this.board[i][j] = new Piece(i, j, 'D', k, null);
-
-                        k++;
-                    }
+                    k++;
                 }
             }
-        } else {
-
-            this.board = state;
         }
     }
 
@@ -46,7 +40,7 @@ export class Board {
     takePiece(piece, currPos, yDiff, nextPos) {
 
         let isTaken = false;
-/*
+
         if(piece.side == 'U' && piece.stack == 1) {
 
             let xPiece = currPos.xPos + 1;
@@ -57,7 +51,7 @@ export class Board {
             if(yDiff > 0)
                 yPiece = currPos.yPos + 1;
 
-            if(this.board[xPiece][yPiece] != null && this.board[xPiece][yPiece].side != piece.side) {
+            if(this.board[xPiece][yPiece] != null) {
                 this.board[xPiece][yPiece] = null;
                 isTaken = true;
             }
@@ -73,14 +67,14 @@ export class Board {
             if(yDiff > 0)
                 yPiece = currPos.yPos - 1;
 
-            if(this.board[xPiece][yPiece] != null && this.board[xPiece][yPiece].side != piece.side) {
+            if(this.board[xPiece][yPiece] != null) {
                 this.board[xPiece][yPiece] = null;
                 isTaken = true;
             }
         }
 
-        if(piece.stack > 1) {
-*/
+        if(piece.stack > 2) {
+
             let xPiece = null;
             let yPiece = null;
 
@@ -89,7 +83,7 @@ export class Board {
                 xPiece = currPos.xPos - 1;
                 yPiece = currPos.yPos - 1;
 
-                if(this.board[xPiece][yPiece] != null && this.board[xPiece][yPiece].side != piece.side) {
+                if(this.board[xPiece][yPiece] != null) {
                     this.board[xPiece][yPiece] = null;
                     isTaken = true;
                 }
@@ -100,7 +94,7 @@ export class Board {
                 xPiece = currPos.xPos - 1;
                 yPiece = currPos.yPos + 1;
 
-                if(this.board[xPiece][yPiece] != null && this.board[xPiece][yPiece].side != piece.side) {
+                if(this.board[xPiece][yPiece] != null) {
                     this.board[xPiece][yPiece] = null;
                     isTaken = true;
                 }
@@ -111,7 +105,7 @@ export class Board {
                 xPiece = currPos.xPos + 1;
                 yPiece = currPos.yPos - 1;
 
-                if(this.board[xPiece][yPiece] != null && this.board[xPiece][yPiece].side != piece.side) {
+                if(this.board[xPiece][yPiece] != null) {
                     this.board[xPiece][yPiece] = null;
                     isTaken = true;
                 }
@@ -122,12 +116,12 @@ export class Board {
                 xPiece = currPos.xPos + 1;
                 yPiece = currPos.yPos + 1;
 
-                if(this.board[xPiece][yPiece] != null && this.board[xPiece][yPiece].side != piece.side) {
+                if(this.board[xPiece][yPiece] != null) {
                     this.board[xPiece][yPiece] = null;
                     isTaken = true;
                 }
             }
-        //}
+        }
 
         return isTaken;
     }
@@ -149,7 +143,7 @@ export class Board {
 
                 let oneSq = (yDiff == 1 || yDiff == -1) && xDiff == 1;
 
-                let twoSq = (xDiff == 2 || xDiff == -2) && (yDiff == 2 || yDiff == -2);
+                let twoSq = (yDiff == 2 || yDiff == -2) && xDiff == 2;
 
                 if(oneSq)
                     legal = true;
@@ -189,7 +183,7 @@ export class Board {
 
                 //console.log(oneSq);
 
-                let twoSq = (xDiff == 2 || xDiff == -2) && (yDiff == 2 || yDiff == -2);
+                let twoSq = (yDiff == 2 || yDiff == -2) && xDiff == 2;
 
                 if(oneSq) {
                     //console.log(nextPos.xPos + ", " + nextPos.yPos);
@@ -223,19 +217,12 @@ export class Board {
 
         if(this.isMoveLegal(piece, nextPos)) {
 
-            this.scanBoard(piece, nextPos);
+            if(nextPos.xPos == 0 || nextPos.xPos == 7)
+                piece.incrementStack();
 
-            let newPiece = new Piece(nextPos.xPos, nextPos.yPos, piece.side, piece.id, piece.stack);
+            this.board[nextPos.xPos][nextPos.yPos] = new Piece(nextPos.xPos, nextPos.yPos, piece.side, piece.id);
 
-            if(nextPos.xPos == 0 && newPiece.side == "D" && newPiece.stack == 1) 
-                newPiece.incrementStack();
-
-            if(nextPos.xPos == 7 && newPiece.side == "U" && newPiece.stack == 1) 
-                newPiece.incrementStack();
-
-            this.board[nextPos.xPos][nextPos.yPos] = newPiece;
-            
-            let currPos = piece.getPosition(); 
+            let currPos = piece.getPosition();
 
             this.board[currPos.xPos][currPos.yPos] = null;
 
@@ -245,73 +232,6 @@ export class Board {
         return moved;
     } 
 
-    scanBoard(piece, nextPos) {
-
-        let i, j;
-
-        for(i = 0; i < 8; i++) {
-            for(j = 0; j < 8; j++) {
-                if(this.board[i][j] != null && this.board[i][j].id != piece.id && this.board[i][j].side == piece.side) {
-                    if(this.checkPiece(this.board[i][j], piece, nextPos)) {
-                        break;
-                    }
-                }
-            }
-        }
-    }
-
-
-    checkPiece(currPiece, piece, nextPos) {
-
-        //console.log(currPiece.id);
-
-        let autoRemove = false;
-
-        let xNext = nextPos.xPos;
-        let yNext = nextPos.yPos;
-
-        let xPos = currPiece.getPosition().xPos;
-        let yPos = currPiece.getPosition().yPos;
-
-        if(0 <= xPos - 2 && 0 <= yPos - 2 && xPos - 2 <= 7 && yPos - 2 <= 7) {
-            //console.log("Upper Left");
-            if(this.board[xPos - 1][yPos - 1] != null && this.board[xPos - 1][yPos - 1].side != piece.side && this.board[xPos - 2][yPos - 2] == null) {
-                this.board[xPos][yPos] = null;
-                autoRemove = true;
-                //console.log("Upper Left");
-            }
-        }
-
-        if(0 <= xPos + 2 && 0 <= yPos + 2 && xPos + 2 <= 7 && yPos + 2 <= 7) {
-            //console.log("Lower Right");
-            if(autoRemove == false && this.board[xPos + 1][yPos + 1] != null && this.board[xPos + 1][yPos + 1].side != piece.side && this.board[xPos + 2][yPos + 2] == null) {
-                this.board[xPos][yPos] = null;
-                autoRemove = true;
-                //console.log("Lower Right");
-            }
-        }
-
-        if(0 <= xPos - 2 && 0 <= yPos + 2 && xPos - 2 <= 7 && yPos + 2 <= 7) {
-            //console.log("Upper Right");
-            if(autoRemove == false && this.board[xPos - 1][yPos + 1] != null && this.board[xPos - 1][yPos + 1].side != piece.side && this.board[xPos - 2][yPos + 2] == null) {
-                this.board[xPos][yPos] = null;
-                autoRemove = true;
-                //console.log("Upper Right");
-            }
-        }
-
-        if(0 <= xPos + 2 && 0 <= yPos - 2 && xPos + 2 <= 7 && yPos - 2 <= 7) {
-            //console.log("Lower Left");
-            if(autoRemove == false && this.board[xPos + 1][yPos - 1] != null && this.board[xPos + 1][yPos - 1].side != piece.side && this.board[xPos + 2][yPos - 2] == null) {
-                this.board[xPos][yPos] = null;
-                autoRemove = true;
-                //console.log("Lower Left");
-            }
-        }
-
-        return autoRemove;
-    }
-
 
     isEmpty(xpos, ypos) {
 
@@ -319,6 +239,7 @@ export class Board {
             return false;
         else
             return true;
+
     }
 
     getId(i, j) {
@@ -335,10 +256,6 @@ export class Board {
         if(this.board[i][j] != null)
             return this.board[i][j];
 
-    }
-
-    getBoard() {
-        return this.board;
     }
 
 }
